Prepare ASN1 marshaling to allow moving to applicationpb
