+++ service/coordinator/coordinator_test.go
@@ -30,6 +30,7 @@ import (
 	"github.com/hyperledger/fabric-x-committer/utils/connection"
 	"github.com/hyperledger/fabric-x-committer/utils/monitoring"
 	"github.com/hyperledger/fabric-x-committer/utils/test"
+	"github.com/hyperledger/fabric-x-committer/utils/test/apptest"
 )
 
 type (
@@ -416,7 +417,7 @@ func TestCoordinatorServiceDependentOrderedTxs(t *testing.T) {
 						ReadWrites: []*applicationpb.ReadWrite{{
 							Key:     mainKey,
 							Value:   []byte("value of version 1"),
-							Version: committerpb.Version(0),
+							Version: applicationpb.NewVersion(0),
 						}},
 					}},
 				},
@@ -442,7 +443,7 @@ func TestCoordinatorServiceDependentOrderedTxs(t *testing.T) {
 						NsVersion: utNsVersion,
 						ReadsOnly: []*applicationpb.Read{{
 							Key:     mainKey,
-							Version: committerpb.Version(2),
+							Version: applicationpb.NewVersion(2),
 						}},
 						ReadWrites: []*applicationpb.ReadWrite{{
 							Key:   subKey,
@@ -459,7 +460,7 @@ func TestCoordinatorServiceDependentOrderedTxs(t *testing.T) {
 						NsVersion: utNsVersion,
 						ReadWrites: []*applicationpb.ReadWrite{{
 							Key:     mainKey,
-							Version: committerpb.Version(2),
+							Version: applicationpb.NewVersion(2),
 							Value:   []byte("Value of version 3"),
 						}},
 					}},
@@ -468,7 +469,7 @@ func TestCoordinatorServiceDependentOrderedTxs(t *testing.T) {
 		},
 	}
 	for _, tx := range b1.Txs {
-		tx.Content.Endorsements = test.CreateEndorsementsForThresholdRule([]byte("dummy"))
+		tx.Content.Endorsements = apptest.CreateEndorsementsForThresholdRule([]byte("dummy"))
 	}
 
 	expectedReceived := test.GetIntMetricValue(t, env.coordinator.metrics.transactionReceivedTotal) + len(b1.Txs)
@@ -679,7 +680,7 @@ func TestCoordinatorRecovery(t *testing.T) {
 							Key: []byte("key3"),
 						}},
 					}},
-					Endorsements: test.CreateEndorsementsForThresholdRule([]byte("dummy")),
+					Endorsements: apptest.CreateEndorsementsForThresholdRule([]byte("dummy")),
 				},
 			},
 			{
@@ -760,7 +761,7 @@ func TestCoordinatorStreamFailureWithSidecar(t *testing.T) {
 							Key: []byte("key1"),
 						}},
 					}},
-					Endorsements: test.CreateEndorsementsForThresholdRule([]byte("dummy")),
+					Endorsements: apptest.CreateEndorsementsForThresholdRule([]byte("dummy")),
 				},
 			},
 		},
@@ -811,7 +812,7 @@ func (e *coordinatorTestEnv) requireStatus(
 	}
 
 	maps.Copy(expectedTxStatus, differentPersisted)
-	test.EnsurePersistedTxStatus(ctx, t, e.client, txIDs, expectedTxStatus)
+	apptest.EnsurePersistedTxStatus(ctx, t, e.client, txIDs, expectedTxStatus)
 }
 
 func (e *coordinatorTestEnv) receiveStatus(t *testing.T, count int) map[string]*applicationpb.StatusWithHeight {
@@ -968,7 +969,7 @@ func makeTestBlock(txPerBlock int) (
 						Key: []byte("key" + strconv.Itoa(i)),
 					}},
 				}},
-				Endorsements: test.CreateEndorsementsForThresholdRule([]byte("dummy")),
+				Endorsements: apptest.CreateEndorsementsForThresholdRule([]byte("dummy")),
 			},
 		}
 		//nolint: gosec // int -> uint32.
