+++ service/vc/validator_committer_service_test.go
@@ -25,6 +25,7 @@ import (
 	"github.com/hyperledger/fabric-x-committer/utils/connection"
 	"github.com/hyperledger/fabric-x-committer/utils/grpcerror"
 	"github.com/hyperledger/fabric-x-committer/utils/test"
+	"github.com/hyperledger/fabric-x-committer/utils/test/apptest"
 )
 
 type validatorAndCommitterServiceTestEnvWithClient struct {
@@ -286,7 +287,7 @@ func TestValidatorAndCommitterService(t *testing.T) {
 								{
 									Key:     []byte("Existing key"),
 									Value:   []byte("new-value"),
-									Version: committerpb.Version(0),
+									Version: applicationpb.NewVersion(0),
 								},
 							},
 						},
@@ -316,7 +317,7 @@ func TestValidatorAndCommitterService(t *testing.T) {
 		env.dbEnv.StatusExistsForNonDuplicateTxID(t, expectedTxStatus)
 
 		ctx, _ := createContext(t)
-		test.EnsurePersistedTxStatus(ctx, t, env.commonClient, txIDs, expectedTxStatus)
+		apptest.EnsurePersistedTxStatus(ctx, t, env.commonClient, txIDs, expectedTxStatus)
 
 		txBatch = &servicepb.VcBatch{
 			Transactions: []*servicepb.VcTx{
@@ -669,7 +670,7 @@ func TestTransactionResubmission(t *testing.T) {
 			env.dbEnv.StatusExistsForNonDuplicateTxID(t, expectedTxStatus)
 		}
 
-		test.EnsurePersistedTxStatus(ctx, t, env.commonClient, txIDs, expectedTxStatus)
+		apptest.EnsurePersistedTxStatus(ctx, t, env.commonClient, txIDs, expectedTxStatus)
 	})
 
 	t.Run("same transactions submitted again while previous submission is not yet committed", func(t *testing.T) {
@@ -685,7 +686,7 @@ func TestTransactionResubmission(t *testing.T) {
 			require.Equal(t, expectedTxStatus, txStatus.Status)
 		}
 		env.dbEnv.StatusExistsForNonDuplicateTxID(t, expectedTxStatus)
-		test.EnsurePersistedTxStatus(ctx, t, env.commonClient, txIDs, expectedTxStatus)
+		apptest.EnsurePersistedTxStatus(ctx, t, env.commonClient, txIDs, expectedTxStatus)
 	})
 
 	t.Run("same transactions submitted again within the minbatchsize", func(t *testing.T) {
@@ -701,7 +702,7 @@ func TestTransactionResubmission(t *testing.T) {
 		require.Equal(t, expectedTxStatus, txStatus.Status)
 
 		env.dbEnv.StatusExistsForNonDuplicateTxID(t, expectedTxStatus)
-		test.EnsurePersistedTxStatus(ctx, t, env.commonClient, txIDs, expectedTxStatus)
+		apptest.EnsurePersistedTxStatus(ctx, t, env.commonClient, txIDs, expectedTxStatus)
 	})
 
 	t.Run("same duplicate transactions submitted in parallel to all vcservices", func(t *testing.T) {
@@ -723,7 +724,7 @@ func TestTransactionResubmission(t *testing.T) {
 			require.Equal(t, expectedTxStatus, txStatus.Status)
 		}
 		env.dbEnv.StatusExistsForNonDuplicateTxID(t, expectedTxStatus)
-		test.EnsurePersistedTxStatus(ctx, t, env.commonClient, txIDs, expectedTxStatus)
+		apptest.EnsurePersistedTxStatus(ctx, t, env.commonClient, txIDs, expectedTxStatus)
 	})
 }
