+++ service/vc/validator_test.go
@@ -63,6 +63,9 @@ func TestValidate(t *testing.T) { //nolint:maintidx // cannot improve.
 	k2_4 := []byte("key2.4")
 	k2_5 := []byte("key2.5")
 
+	// Shortcut for version pointer.
+	v := applicationpb.NewVersion
+
 	env.dbEnv.populateData(
 		t,
 		[]string{"1", "2"},
@@ -130,26 +133,26 @@ func TestValidate(t *testing.T) { //nolint:maintidx // cannot improve.
 				nsToReads: namespaceToReads{
 					"1": &reads{
 						keys:     [][]byte{k1_1, k1_2, k1_5},
-						versions: []*uint64{committerpb.Version(1), committerpb.Version(1), nil},
+						versions: []*uint64{v(1), v(1), nil},
 					},
 					"2": &reads{
 						keys:     [][]byte{k2_1, k2_2, k2_5},
-						versions: []*uint64{committerpb.Version(0), committerpb.Version(0), nil},
+						versions: []*uint64{v(0), v(0), nil},
 					},
 					committerpb.MetaNamespaceID: &reads{
 						keys:     [][]byte{[]byte("1"), []byte("2")},
-						versions: []*uint64{committerpb.Version(0), committerpb.Version(0)},
+						versions: []*uint64{v(0), v(0)},
 					},
 				},
 				readToTxIDs: readToTransactions{
-					newCmpRead("1", k1_1, committerpb.Version(1)):                                []TxID{"tx1"},
-					newCmpRead("1", k1_2, committerpb.Version(1)):                                []TxID{"tx1"},
-					newCmpRead("1", k1_5, nil):                                                   []TxID{"tx2"},
-					newCmpRead("2", k2_1, committerpb.Version(0)):                                []TxID{"tx1"},
-					newCmpRead("2", k2_2, committerpb.Version(0)):                                []TxID{"tx3"},
-					newCmpRead("2", k2_5, nil):                                                   []TxID{"tx3"},
-					newCmpRead(committerpb.MetaNamespaceID, []byte("1"), committerpb.Version(0)): []TxID{"tx1", "tx2"},
-					newCmpRead(committerpb.MetaNamespaceID, []byte("2"), committerpb.Version(0)): []TxID{"tx1", "tx3"},
+					newCmpRead("1", k1_1, v(1)):                                []TxID{"tx1"},
+					newCmpRead("1", k1_2, v(1)):                                []TxID{"tx1"},
+					newCmpRead("1", k1_5, nil):                                 []TxID{"tx2"},
+					newCmpRead("2", k2_1, v(0)):                                []TxID{"tx1"},
+					newCmpRead("2", k2_2, v(0)):                                []TxID{"tx3"},
+					newCmpRead("2", k2_5, nil):                                 []TxID{"tx3"},
+					newCmpRead(committerpb.MetaNamespaceID, []byte("1"), v(0)): []TxID{"tx1", "tx2"},
+					newCmpRead(committerpb.MetaNamespaceID, []byte("2"), v(0)): []TxID{"tx1", "tx3"},
 				},
 				txIDToNsNonBlindWrites: transactionToWrites{
 					"tx1": tx1NonBlindWrites,
@@ -189,7 +192,7 @@ func TestValidate(t *testing.T) { //nolint:maintidx // cannot improve.
 				nsToReads: namespaceToReads{
 					"1": &reads{
 						keys:     [][]byte{k1_1, k1_2, k1_5},
-						versions: []*uint64{committerpb.Version(0), committerpb.Version(0), committerpb.Version(1)},
+						versions: []*uint64{v(0), v(0), v(1)},
 					},
 					"2": &reads{
 						keys:     [][]byte{k2_1, k2_2, k2_5},
@@ -197,18 +200,18 @@ func TestValidate(t *testing.T) { //nolint:maintidx // cannot improve.
 					},
 					committerpb.MetaNamespaceID: &reads{
 						keys:     [][]byte{[]byte("1"), []byte("2")},
-						versions: []*uint64{committerpb.Version(1), committerpb.Version(1)},
+						versions: []*uint64{v(1), v(1)},
 					},
 				},
 				readToTxIDs: readToTransactions{
-					newCmpRead("1", k1_1, committerpb.Version(0)):                                []TxID{"tx1"},
-					newCmpRead("1", k1_2, committerpb.Version(0)):                                []TxID{"tx1"},
-					newCmpRead("1", k1_5, committerpb.Version(1)):                                []TxID{"tx2"},
-					newCmpRead("2", k2_1, nil):                                                   []TxID{"tx1"},
-					newCmpRead("2", k2_2, nil):                                                   []TxID{"tx3"},
-					newCmpRead("2", k2_5, nil):                                                   []TxID{"tx3"},
-					newCmpRead(committerpb.MetaNamespaceID, []byte("1"), committerpb.Version(1)): []TxID{"tx1", "tx2"},
-					newCmpRead(committerpb.MetaNamespaceID, []byte("2"), committerpb.Version(1)): []TxID{"tx1", "tx2"},
+					newCmpRead("1", k1_1, v(0)):                                []TxID{"tx1"},
+					newCmpRead("1", k1_2, v(0)):                                []TxID{"tx1"},
+					newCmpRead("1", k1_5, v(1)):                                []TxID{"tx2"},
+					newCmpRead("2", k2_1, nil):                                 []TxID{"tx1"},
+					newCmpRead("2", k2_2, nil):                                 []TxID{"tx3"},
+					newCmpRead("2", k2_5, nil):                                 []TxID{"tx3"},
+					newCmpRead(committerpb.MetaNamespaceID, []byte("1"), v(1)): []TxID{"tx1", "tx2"},
+					newCmpRead(committerpb.MetaNamespaceID, []byte("2"), v(1)): []TxID{"tx1", "tx2"},
 				},
 				txIDToNsNonBlindWrites: transactionToWrites{
 					"tx1": tx1NonBlindWrites,
@@ -246,7 +249,7 @@ func TestValidate(t *testing.T) { //nolint:maintidx // cannot improve.
 				nsToReads: namespaceToReads{
 					"1": &reads{
 						keys:     [][]byte{k1_1, k1_2, k1_5},
-						versions: []*uint64{committerpb.Version(1), committerpb.Version(1), nil},
+						versions: []*uint64{v(1), v(1), nil},
 					},
 					"2": &reads{
 						keys:     [][]byte{k2_1, k2_2, k2_5},
@@ -258,19 +261,19 @@ func TestValidate(t *testing.T) { //nolint:maintidx // cannot improve.
 							[]byte("2"),
 							[]byte("2"),
 						},
-						versions: []*uint64{committerpb.Version(0), committerpb.Version(0), committerpb.Version(1)},
+						versions: []*uint64{v(0), v(0), v(1)},
 					},
 				},
 				readToTxIDs: readToTransactions{
-					newCmpRead("1", k1_1, committerpb.Version(1)):                                []TxID{"tx1"},
-					newCmpRead("1", k1_2, committerpb.Version(1)):                                []TxID{"tx1"},
-					newCmpRead("1", k1_5, nil):                                                   []TxID{"tx2"},
-					newCmpRead("2", k2_1, nil):                                                   []TxID{"tx1"},
-					newCmpRead("2", k2_2, nil):                                                   []TxID{"tx3"},
-					newCmpRead("2", k2_5, nil):                                                   []TxID{"tx3"},
-					newCmpRead(committerpb.MetaNamespaceID, []byte("1"), committerpb.Version(0)): []TxID{"tx1", "tx2"},
-					newCmpRead(committerpb.MetaNamespaceID, []byte("2"), committerpb.Version(0)): []TxID{"tx1", "tx3"},
-					newCmpRead(committerpb.MetaNamespaceID, []byte("2"), committerpb.Version(1)): []TxID{"tx4"},
+					newCmpRead("1", k1_1, v(1)):                                []TxID{"tx1"},
+					newCmpRead("1", k1_2, v(1)):                                []TxID{"tx1"},
+					newCmpRead("1", k1_5, nil):                                 []TxID{"tx2"},
+					newCmpRead("2", k2_1, nil):                                 []TxID{"tx1"},
+					newCmpRead("2", k2_2, nil):                                 []TxID{"tx3"},
+					newCmpRead("2", k2_5, nil):                                 []TxID{"tx3"},
+					newCmpRead(committerpb.MetaNamespaceID, []byte("1"), v(0)): []TxID{"tx1", "tx2"},
+					newCmpRead(committerpb.MetaNamespaceID, []byte("2"), v(0)): []TxID{"tx1", "tx3"},
+					newCmpRead(committerpb.MetaNamespaceID, []byte("2"), v(1)): []TxID{"tx4"},
 				},
 				txIDToNsNonBlindWrites: transactionToWrites{
 					"tx1": tx1NonBlindWrites,
